Clean up constants file
