+++ packages/cli/src/modules/insights/__tests__/insights.service.test.ts
@@ -26,7 +26,6 @@ import type { InsightsRaw } from '../database/entities/insights-raw';
 import type { InsightsByPeriodRepository } from '../database/repositories/insights-by-period.repository';
 import { InsightsCollectionService } from '../insights-collection.service';
 import { InsightsCompactionService } from '../insights-compaction.service';
-import { getAvailableDateRanges } from '../insights-helpers';
 import type { InsightsPruningService } from '../insights-pruning.service';
 import { InsightsConfig } from '../insights.config';
 import { InsightsService } from '../insights.service';
@@ -584,16 +583,26 @@ describe('getInsightsByTime', () => {
 
 describe('getAvailableDateRanges', () => {
 	let licenseMock: jest.Mocked<LicenseState>;
+	let insightsService: InsightsService;
 
 	beforeAll(() => {
 		licenseMock = mock<LicenseState>();
+		insightsService = new InsightsService(
+			mock(),
+			mock(),
+			mock(),
+			mock(),
+			licenseMock,
+			mock(),
+			mockLogger(),
+		);
 	});
 
 	test('returns correct ranges when hourly data is enabled and max history is unlimited', () => {
 		licenseMock.getInsightsMaxHistory.mockReturnValue(-1);
 		licenseMock.isInsightsHourlyDataLicensed.mockReturnValue(true);
 
-		const result = getAvailableDateRanges(licenseMock);
+		const result = insightsService.getAvailableDateRanges();
 
 		expect(result).toEqual([
 			{ key: 'day', licensed: true, granularity: 'hour' },
@@ -610,7 +619,7 @@ describe('getAvailableDateRanges', () => {
 		licenseMock.getInsightsMaxHistory.mockReturnValue(365);
 		licenseMock.isInsightsHourlyDataLicensed.mockReturnValue(true);
 
-		const result = getAvailableDateRanges(licenseMock);
+		const result = insightsService.getAvailableDateRanges();
 
 		expect(result).toEqual([
 			{ key: 'day', licensed: true, granularity: 'hour' },
@@ -627,7 +636,7 @@ describe('getAvailableDateRanges', () => {
 		licenseMock.getInsightsMaxHistory.mockReturnValue(30);
 		licenseMock.isInsightsHourlyDataLicensed.mockReturnValue(false);
 
-		const result = getAvailableDateRanges(licenseMock);
+		const result = insightsService.getAvailableDateRanges();
 
 		expect(result).toEqual([
 			{ key: 'day', licensed: false, granularity: 'hour' },
@@ -644,7 +653,7 @@ describe('getAvailableDateRanges', () => {
 		licenseMock.getInsightsMaxHistory.mockReturnValue(5);
 		licenseMock.isInsightsHourlyDataLicensed.mockReturnValue(false);
 
-		const result = getAvailableDateRanges(licenseMock);
+		const result = insightsService.getAvailableDateRanges();
 
 		expect(result).toEqual([
 			{ key: 'day', licensed: false, granularity: 'hour' },
@@ -661,7 +670,7 @@ describe('getAvailableDateRanges', () => {
 		licenseMock.getInsightsMaxHistory.mockReturnValue(90);
 		licenseMock.isInsightsHourlyDataLicensed.mockReturnValue(true);
 
-		const result = getAvailableDateRanges(licenseMock);
+		const result = insightsService.getAvailableDateRanges();
 
 		expect(result).toEqual([
 			{ key: 'day', licensed: true, granularity: 'hour' },
